[IMP] mergebot: allow running tests without verbose github logging
Mergebot & forwardbot have ultra-verbose logging of all github
interactions in order to better understand what happens exactly when
there are issues with gh integration (and/or provide to GH support).

However in most cases this is a pain in the ass when reviewing test
logs. So suppress these github_requests logs by default when testing.