[IMP] runbot: use github pull requests labels
In order to facilitate some filters and searches, this commit allows to
retrieve PR labels from Github. They are linked to branches and a bundle
will hold the labels from all its PR's.

They can be searched as it's the final purpose of this PR.

Although this could be done on Github, the runbot aggregates branches
from multiple repos in a bundle. So it's easier to browse and search
bundles than doing the same in multiple Github tabs.