[IMP] runbot: change total build time
The build time was changed in order to give the time between the start
and the end of all subbuild. Unfortunately this info does not give a
real idea of a build time since it will depends on load, pendings, ...

This commit gos back to a more simple logic where the build time is
between the start and the endof the build, without children. A total
time can be computed based on children if needed, giving the total
computation time for all builds and children.