[FIX] runbot: better base management
The way branch_info, bundle_id and base_id are computed have multiple
issue:
-> branch_info is a compute called manually sometimes, and difficult
to define as a compute since it is based on external data.
Making a setup function for that makes its management much easier. Call
it when needed
-> creating a bundle in a compute is a bad idea. This should be done
before. Making the compute a setup function only called on creation
makes it easier to manage. A bundle shouldn't change for a branch exept
manually.
-> a hack is used to define a "defined_base_id" when the pr changes
target, or if the target does not match the name. The target should be
enough when a pr exists. This can lead to error when opening a new pr
with the same branch name. The last alive pr of a batch should be the
one to use in most cases. This can be taken into account inside the
_compute_base_id.

This commit also rework how a trigger is started in order to avoid
starting some build if there is no pr alive, this will be usefull
for style, security, and maybe upgrades.