[IMP] runbot: better docker start management
The flush was initially usefull to ensure that the logcounter is correct
in database before starting the docker when logs wher using a
postgresql trigger. This is not usefull anymore.

This flush as a side effect was also ensuring that everything was
was flushe before starting the docker. It would even be better to commit
Stating the docker after the commit is event better and safer in this.

The docker run is independant from the database and is returned to the
scheduler loop to be started after the transaction is commited.