[IMP] use queue for global status
Writing a global_state every time creates concurrent updates:

- 10 subbuild are going from `pending` to `testing`
- 10 concurrent write on the parent is "waiting", goes "waiting"
- 9! potential transaction rollbacks

The initial proposed solution is to avoid writing if the state doesn't
changes:

- 10 subbuild are going from `pending` to `testing`
- 10 concurrent write avoided: parent state didn't changed
- at most one transaction rollback per state changed

But this solution doesn't actually ensures consistency: works only if a
state can be computed based on it own state + current parent state
- parent is ok and I am ko -> parent is ko
- parent is ko and I am ko -> parent is ko (no change, it's ok)

This is not True for state

- parent is waiting and I am done -> depends on other child state.
This means that to compute the global_state, it is needed to read
children.

In repeatable read, if any other child changes of state, the transaction
won't be rollbacked in repeatable read. Meaning that if all build are
going from 'testing' to 'done' at the same time, the parent will stay
in waiting (concurrent transaction are not aware of other child changes)