adapt host docker_build for using a docker registry
If the host is the docker registry used globally, it should build the
docker image. Otherwise, if the host is not using a registry to pull the
images, it should build them locally.

In either cases, it should push the local images to the local registry
as each runbot builder is hosting a Docker registry.