[IMP] runbot: search build errors in build error logs
When managing build errors, it's sometimes useful to search for other
occurrences of the error in all the build logs. Although it's possible
to do it manually, the operation is tedious.

With this commit, a button is added on the error page to automatically
perform the search for the same error in all the build error logs
(sql view).