[IMP] runbot: add qualifiers on build error
This commit adds a `qualifiers` field on build error. Its purpose is to
find similary qualified errors based on the qualified error contents.

A new tab is added on the build error form. When qualifiers are added on
the error, the similar errors are computed by searching errors having
their content similary qualifieds. When found, a button can be used to
merge those errors with the current one.

Also, a button to suggest qualifiers is added. This buttons suggest
qualifiers by assembling common qualifiers on the error content's.