[IMP] runbot: introduce matrix for upgrade
The initial way to configure upgrades was working fine and automatically
but does not alway match the reality of what should be supported.

Saying, upgrade from the last main version and last intermediate
(for 18, from 17.0 and 17.4) we may also want to test from 17.2 because
for some reason there are many people in 17.2 taht will need to upgrade
in 18.0. But the transition 17.2->17.3 doesn't make sense anymore
after a while since people will go in 18 imediatly at some point.

This matrix can be generated automatically from the parameters and new
versions should appear automatically, but it is possible to tweak the
values on demand.