[IMP] runbot: add support for dockerfile variants
Dockerfile variants are dockerfile that depend on another dockerfile
while adding layers, which should not be present in the regular image.
Strictly speaking it is the same as creating a new dockerfile with a
reference docker layer at the top, except that layer is implicit and
will not clutter the docker image.

Variants must be unique.