[FIX] runbot: fix rmtree permission issues
When trying to garbage collect directories if a file or a directory
is not writable for the owner, the garbage collecting fails.

The worst case is when both a file and its parent directory are not
writable by the user. In that case, even by changing the file
permission, the runbot cannot remove the file because of the directory
not being writable.

With this commit, a callback function is used that tries to change the
owner permission of the file and its parent directory to be writable
before retrying to unlink them.