[FIX] runbot: fix rmtree permission issues
When trying to garbaged collect diretories if a file or a directory
that the runbot user process own is not writable for the owner, the
garbage collecting fails.

The worst case is when both a file and its parent directory are not
writable by the user. In that case, even by changing the file
permission, the runbot cannot remove the file because of the directory
not being writable.

With this commit, a new `rmforest` function is made that tries, in case
of permission error, to change the owner permission of the file and its
parent directory to being writable before retrying to unlink them.